requirements.txt
# Core Dependencies
fastapi~=0.115.8
SQLAlchemy==2.0.38
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2

# Testing and Code Quality
pytest==8.3.4
pytest-cov==6.0.0
pylint==3.3.2

# Development Tools
python-dotenv==1.0.1
black==24.10.0
isort==5.13.2

# Database Migrations
alembic==1.14.1

# API Client and Server
httpx==0.28.1
uvicorn==0.34.0